requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
libsql-client>=0.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.4.2
email-validator>=2.0.0